Refactor publication service classes
