+++ src/integrationTest/java/uk/gov/hmcts/reform/pip/data/management/controllers/publication/PublicationManagementTest.java
@@ -54,7 +54,7 @@ import static uk.gov.hmcts.reform.pip.model.publication.FileType.PDF;
 @AutoConfigureEmbeddedDatabase(type = AutoConfigureEmbeddedDatabase.DatabaseType.POSTGRES)
 @WithMockUser(username = "admin", authorities = {"APPROLE_api.request.admin"})
 @SuppressWarnings({"PMD.ExcessiveImports"})
-class PublicationFileManagementTest extends PublicationIntegrationTestBase {
+class PublicationManagementTest extends PublicationIntegrationTestBase {
     private static final String ROOT_URL = "/publication";
     private static final String GET_FILE_URL = ROOT_URL + "/%s/%s";
     private static final String ARTEFACT_ID = UUID.randomUUID().toString();
