+++ src/integrationTest/java/uk/gov/hmcts/reform/pip/data/management/service/PublicationFileGenerationServiceTest.java
@@ -14,7 +14,8 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import uk.gov.hmcts.reform.pip.data.management.models.PublicationFiles;
 import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
-import uk.gov.hmcts.reform.pip.data.management.service.publication.ArtefactService;
+import uk.gov.hmcts.reform.pip.data.management.service.location.LocationService;
+import uk.gov.hmcts.reform.pip.data.management.service.publication.PublicationRetrievalService;
 import uk.gov.hmcts.reform.pip.data.management.utils.IntegrationBasicTestBase;
 import uk.gov.hmcts.reform.pip.model.publication.Language;
 import uk.gov.hmcts.reform.pip.model.publication.ListType;
@@ -54,7 +55,7 @@ class PublicationFileGenerationServiceTest extends IntegrationBasicTestBase {
     private String civilDailyListInput;
 
     @MockitoBean
-    private ArtefactService artefactService;
+    private PublicationRetrievalService artefactService;
 
     @MockitoBean
     private LocationService locationService;
