+++ src/main/java/uk/gov/hmcts/reform/pip/data/management/service/publication/PublicationCreationRunner.java
@@ -20,16 +20,16 @@ import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
 @Slf4j
 @Service
 public class PublicationCreationRunner {
-    private final PublicationService publicationService;
+    private final PublicationCreationService publicationCreationService;
 
-    private final ArtefactService artefactService;
+    private final PublicationRetrievalService artefactService;
 
     private final JsonExtractor jsonExtractor;
 
     @Autowired
-    public PublicationCreationRunner(PublicationService publicationService, ArtefactService artefactService,
+    public PublicationCreationRunner(PublicationCreationService publicationCreationService, PublicationRetrievalService artefactService,
                                      JsonExtractor jsonExtractor) {
-        this.publicationService = publicationService;
+        this.publicationCreationService = publicationCreationService;
         this.artefactService = artefactService;
         this.jsonExtractor = jsonExtractor;
     }
@@ -47,7 +47,7 @@ public class PublicationCreationRunner {
         Artefact createdArtefact;
 
         try {
-            createdArtefact = publicationService.createPublication(artefact, payload);
+            createdArtefact = publicationCreationService.createPublication(artefact, payload);
         } catch (CannotAcquireLockException | DataIntegrityViolationException ex) {
             throw new CreateArtefactConflictException(
                 "Deadlock when creating json publication. Please try again later."
@@ -71,7 +71,7 @@ public class PublicationCreationRunner {
         Artefact createdArtefact;
 
         try {
-            createdArtefact = publicationService.createPublication(artefact, file);
+            createdArtefact = publicationCreationService.createPublication(artefact, file);
         } catch (CannotAcquireLockException | DataIntegrityViolationException ex) {
             throw new CreateArtefactConflictException(
                 "Deadlock when creating flat file publication. Please try again later."
@@ -95,7 +95,7 @@ public class PublicationCreationRunner {
     }
 
     private void preprocessPublicationForCreation(Artefact artefact) {
-        publicationService.applyInternalLocationId(artefact);
+        publicationCreationService.applyInternalLocationId(artefact);
         artefact.setContentDate(artefact.getContentDate().toLocalDate().atTime(LocalTime.MIN));
         artefact.setLastReceivedDate(LocalDateTime.now());
     }
