+++ src/main/java/uk/gov/hmcts/reform/pip/data/management/service/publication/PublicationDeleteService.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.pip.data.management.service.publication;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,28 +9,21 @@ import uk.gov.hmcts.reform.pip.data.management.database.LocationRepository;
 import uk.gov.hmcts.reform.pip.data.management.errorhandling.exceptions.ArtefactNotFoundException;
 import uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactHelper;
 import uk.gov.hmcts.reform.pip.data.management.helpers.NoMatchArtefactHelper;
-import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
 import uk.gov.hmcts.reform.pip.data.management.service.AccountManagementService;
-import uk.gov.hmcts.reform.pip.data.management.service.LocationService;
+import uk.gov.hmcts.reform.pip.data.management.service.location.LocationService;
 import uk.gov.hmcts.reform.pip.data.management.service.PublicationManagementService;
 import uk.gov.hmcts.reform.pip.data.management.service.PublicationServicesService;
-import uk.gov.hmcts.reform.pip.model.account.PiUser;
 import uk.gov.hmcts.reform.pip.model.enums.UserActions;
-import uk.gov.hmcts.reform.pip.model.system.admin.ActionResult;
-import uk.gov.hmcts.reform.pip.model.system.admin.ChangeType;
 
 import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Stream;
 
 import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
 
 @Slf4j
 @Service
-public class ArtefactDeleteService {
+public class PublicationDeleteService {
 
     private final ArtefactRepository artefactRepository;
     private final LocationRepository locationRepository;
@@ -41,12 +33,12 @@ public class ArtefactDeleteService {
     private final AccountManagementService accountManagementService;
     private final PublicationServicesService publicationServicesService;
 
-    public ArtefactDeleteService(ArtefactRepository artefactRepository, LocationRepository locationRepository,
-                                 LocationService locationService,
-                                 PublicationManagementService publicationManagementService,
-                                 AzureArtefactBlobService azureArtefactBlobService,
-                                 AccountManagementService accountManagementService,
-                                 PublicationServicesService publicationServicesService) {
+    public PublicationDeleteService(ArtefactRepository artefactRepository, LocationRepository locationRepository,
+                                    LocationService locationService,
+                                    PublicationManagementService publicationManagementService,
+                                    AzureArtefactBlobService azureArtefactBlobService,
+                                    AccountManagementService accountManagementService,
+                                    PublicationServicesService publicationServicesService) {
         this.artefactRepository = artefactRepository;
         this.locationRepository = locationRepository;
         this.locationService = locationService;
@@ -120,52 +112,7 @@ public class ArtefactDeleteService {
         log.info(writeLog(issuerId, UserActions.REMOVE, artefactId));
     }
 
-    public String deleteArtefactByLocation(Integer locationId, String userId)
-        throws JsonProcessingException {
-        List<Artefact> activeArtefacts = artefactRepository.findActiveArtefactsForLocation(LocalDateTime.now(),
-                                                                                           locationId.toString());
-        if (activeArtefacts.isEmpty()) {
-            log.info(writeLog(String.format("User %s attempting to delete all artefacts for location %s. "
-                                                + "No artefacts found",
-                                            userId, locationId)));
-            throw new ArtefactNotFoundException(String.format(
-                "No artefacts found with the location ID %s",
-                locationId
-            ));
-        }
-        log.info(writeLog(String.format("User %s attempting to delete all artefacts for location %s. "
-                                            + "%s artefact(s) found",
-                                        userId, locationId, activeArtefacts.size())));
-
-        activeArtefacts.forEach(artefact -> {
-            handleArtefactDeletion(artefact);
-            log.info(writeLog(
-                String.format("Artefact deleted by %s, with artefact id: %s",
-                              userId, artefact.getArtefactId())
-            ));
-        });
-        Optional<Location> location = locationRepository.getLocationByLocationId(locationId);
-        notifySystemAdminAboutSubscriptionDeletion(userId,
-            String.format("Total %s artefact(s) for location %s", activeArtefacts.size(),
-                          location.isPresent() ? location.get().getName() : ""));
-        return String.format("Total %s artefact deleted for location id %s", activeArtefacts.size(), locationId);
-    }
-
-    public String deleteAllArtefactsWithLocationNamePrefix(String prefix) {
-        List<String> locationIds = locationService.getAllLocationsWithNamePrefix(prefix).stream()
-            .map(Object::toString)
-            .toList();
-
-        List<Artefact> artefactsToDelete = Collections.emptyList();
-        if (!locationIds.isEmpty()) {
-            artefactsToDelete = artefactRepository.findAllByLocationIdIn(locationIds);
-            artefactsToDelete.forEach(this::handleArtefactDeletion);
-        }
-        return String.format("%s artefacts(s) deleted for location name starting with %s",
-                             artefactsToDelete.size(), prefix);
-    }
-
-    private void handleArtefactDeletion(Artefact artefact) {
+    public void handleArtefactDeletion(Artefact artefact) {
         deleteDataFromBlobStore(artefact);
         artefactRepository.delete(artefact);
         if (!NoMatchArtefactHelper.isNoMatchLocationId(artefact.getLocationId())) {
@@ -176,15 +123,5 @@ public class ArtefactDeleteService {
     private void handleArtefactArchiving(Artefact artefact) {
         deleteDataFromBlobStore(artefact);
         artefactRepository.archiveArtefact(artefact.getArtefactId().toString());
-    }
-
-    private void notifySystemAdminAboutSubscriptionDeletion(String userId, String additionalDetails)
-        throws JsonProcessingException {
-        PiUser userInfo = accountManagementService.getUserById(userId);
-        List<String> systemAdminsAad = accountManagementService.getAllAccounts("PI_AAD", "SYSTEM_ADMIN");
-        List<String> systemAdminsSso = accountManagementService.getAllAccounts("SSO", "SYSTEM_ADMIN");
-        List<String> systemAdmins = Stream.concat(systemAdminsAad.stream(), systemAdminsSso.stream()).toList();
-        publicationServicesService.sendSystemAdminEmail(systemAdmins, userInfo.getEmail(),
-            ActionResult.SUCCEEDED, additionalDetails, ChangeType.DELETE_LOCATION_ARTEFACT);
     }
 }
