+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/controllers/publication/PublicationSubscriptionControllerTest.java
@@ -6,7 +6,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
-import uk.gov.hmcts.reform.pip.data.management.service.publication.ArtefactTriggerService;
+import uk.gov.hmcts.reform.pip.data.management.service.publication.PublicationSubscriptionService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.doNothing;
@@ -15,14 +15,14 @@ import static uk.gov.hmcts.reform.pip.data.management.helpers.ConstantsTestHelpe
 @ExtendWith(MockitoExtension.class)
 class PublicationSubscriptionControllerTest {
     @Mock
-    private ArtefactTriggerService artefactTriggerService;
+    private PublicationSubscriptionService publicationSubscriptionService;
 
     @InjectMocks
     private PublicationSubscriptionController publicationSubscriptionController;
 
     @Test
     void testSendNewArtefactsForSubscriptionSuccess() {
-        doNothing().when(artefactTriggerService).checkNewlyActiveArtefacts();
+        doNothing().when(publicationSubscriptionService).checkNewlyActiveArtefacts();
         assertThat(publicationSubscriptionController.sendNewArtefactsForSubscription().getStatusCode())
             .as(STATUS_CODE_MATCH)
             .isEqualTo(HttpStatus.NO_CONTENT);
