+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/service/publication/PublicationRetrievalServiceTest.java
@@ -15,14 +15,11 @@ import uk.gov.hmcts.reform.pip.data.management.database.LocationRepository;
 import uk.gov.hmcts.reform.pip.data.management.errorhandling.exceptions.NotFoundException;
 import uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper;
 import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
-import uk.gov.hmcts.reform.pip.data.management.models.location.LocationArtefact;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
 import uk.gov.hmcts.reform.pip.data.management.service.AccountManagementService;
-import uk.gov.hmcts.reform.pip.model.location.LocationType;
 import uk.gov.hmcts.reform.pip.model.publication.ListType;
 import uk.gov.hmcts.reform.pip.model.publication.Sensitivity;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -34,7 +31,6 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.ARTEFACT_ID;
-import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.LOCATION_TYPE_MATCH;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.LOCATION_VENUE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.PROVENANCE;
 import static uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactConstantTestHelper.PROVENANCE_ID;
@@ -50,7 +46,7 @@ import static uk.gov.hmcts.reform.pip.data.management.helpers.ConstantsTestHelpe
 @ActiveProfiles("test")
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("PMD.ExcessiveImports")
-class ArtefactServiceTest {
+class PublicationRetrievalServiceTest {
     private static final String PAYLOAD = "payload";
 
     @Mock
@@ -66,7 +62,7 @@ class ArtefactServiceTest {
     AccountManagementService accountManagementService;
 
     @InjectMocks
-    ArtefactService artefactService;
+    PublicationRetrievalService artefactService;
 
     private Artefact artefact;
     private Artefact artefactClassified;
@@ -325,68 +321,5 @@ class ArtefactServiceTest {
             .thenReturn(Optional.of(artefactWithIdAndPayloadUrl));
         assertEquals(artefactWithIdAndPayloadUrl, artefactService.getMetadataByArtefactId(ARTEFACT_ID, USER_ID),
                      VALIDATION_ARTEFACT_NOT_MATCH);
-    }
-
-    @Test
-    void testGetLocationTypeVenue() {
-        List<ListType> venueListTypes = new ArrayList<>();
-        venueListTypes.add(ListType.CROWN_DAILY_LIST);
-        venueListTypes.add(ListType.CROWN_FIRM_LIST);
-        venueListTypes.add(ListType.CROWN_WARNED_LIST);
-        venueListTypes.add(ListType.MAGISTRATES_PUBLIC_LIST);
-        venueListTypes.add(ListType.MAGISTRATES_STANDARD_LIST);
-        venueListTypes.add(ListType.CIVIL_DAILY_CAUSE_LIST);
-        venueListTypes.add(ListType.FAMILY_DAILY_CAUSE_LIST);
-        venueListTypes.add(ListType.IAC_DAILY_LIST);
-
-        venueListTypes.forEach(listType ->
-                                   assertEquals(
-                                       LocationType.VENUE,
-                                       artefactService.getLocationType(listType),
-                                       LOCATION_TYPE_MATCH));
-    }
-
-    @Test
-    void testGetLocationTypeOwningHearingLocation() {
-        List<ListType> venueListTypes = new ArrayList<>();
-        venueListTypes.add(ListType.ET_DAILY_LIST);
-        venueListTypes.add(ListType.ET_FORTNIGHTLY_PRESS_LIST);
-
-        venueListTypes.forEach(listType ->
-                                   assertEquals(LocationType.OWNING_HEARING_LOCATION,
-                                                artefactService.getLocationType(listType),
-                                                LOCATION_TYPE_MATCH));
-    }
-
-    @Test
-    void testGetLocationTypeNational() {
-        List<ListType> nationalListTypes = new ArrayList<>();
-        nationalListTypes.add(ListType.SJP_PRESS_LIST);
-        nationalListTypes.add(ListType.SJP_PUBLIC_LIST);
-
-        nationalListTypes.forEach(listType ->
-                                      assertEquals(LocationType.NATIONAL,
-                                                   artefactService.getLocationType(listType),
-                                                   LOCATION_TYPE_MATCH));
-
-    }
-
-    @Test
-    void testArtefactCountService() {
-        List<LocationArtefact> artefactsPerLocations = new ArrayList<>();
-        artefactsPerLocations.add(new LocationArtefact("1", 3));
-        artefactsPerLocations.add(new LocationArtefact("noMatch", 0));
-        List<Object[]> result = new ArrayList<>();
-        result.add(new Object[]{"1", "3"});
-        when(artefactRepository.countArtefactsByLocation()).thenReturn(result);
-        assertEquals(artefactsPerLocations, artefactService.countArtefactsByLocation(),
-                     MESSAGES_MATCH);
-    }
-
-    @Test
-    void testFindAllNoMatchArtefacts() {
-        when(artefactRepository.findAllNoMatchArtefacts()).thenReturn(List.of(artefact));
-
-        assertEquals(List.of(artefact), artefactService.findAllNoMatchArtefacts(), MESSAGES_MATCH);
     }
 }
